Technical/Change RecordsDictionaryBuilder namespace (#11)
* Changed RecordsDictionaryBuilder namespace, updated tests
* Added DnsMock::ArgumentTypeError, tests
* Added RecordsDictionaryBuilder argument type checking
